requirements.txt
backcall==0.2.0
decorator==4.4.2
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
jupyter-client==6.1.7
jupyter-core==4.6.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.1
pytesseract==0.3.6
python-dateutil==2.8.1
pyzmq==19.0.2
six==1.15.0
tornado==6.0.4
traitlets==5.0.5
wcwidth==0.2.5
